구현 : Firebase를 이용한 구글 로그인/아웃 함수 만들기 - AuthService 클래스 만들기 - 로그인/아웃 함수 구현 Issue #119
